app/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
prometheus-client==0.20.0
opentelemetry-sdk==1.27.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
structlog==24.4.0
